extend '-js' to allow arbitrary files as kernels
By default the argument is meant to be symbolic of the kernels
linkage and loaded from a specific location within the iocamljs
install (with a bit of name mangling - kernel.<name>.js).

It now allows an arbitrary file to be specified so long as it both
exists and has a '.js' extension.

Of course this was more hassle than expected as it interacts with
html generation, uri->file path handling and static site generation.

Seems to work though, and isn't much messier than before.  I've tested
basic operation and static site generation and it looks to be working
correctly.  As a side note, the whole uri/path handling is getting a
bit out of control and should be simplified (or made more explicit)
at some point.